provide an option to skip uninstall

+++ client/src/main/java/com/terracottatech/qa/angela/client/Tsa.java
@@ -42,6 +42,7 @@ public class Tsa implements AutoCloseable {
 
   private final static Logger logger = LoggerFactory.getLogger(Tsa.class);
   private static final long TIMEOUT = 60000;
+  private static final boolean SKIP_UNINSTALL = Boolean.getBoolean("skipUninstall");
 
   private final Topology topology;
   private final Ignite ignite;
@@ -273,7 +274,9 @@ public class Tsa implements AutoCloseable {
       logger.error("Error when trying to stop servers" , e.getMessage());
       // ignore, not installed
     }
-    uninstallAll();
+    if (!SKIP_UNINSTALL) {
+      uninstallAll();
+    }
   }
